Implement a Log4J2 ContextDataProvider
Closes #4

VertxContextDataProvider supplies logging events with contextual data.

The advantage is that data can then be retrieved anywhere log4j2 supports context map lookups.
The drawback is the provider is invoked anytime a logging event is created (an event is not created if the log level is below the configured threshold).