UOE may be thrown when using log4j2 with context data provider
Fixes #97

VertxContextDataProvider relied on the default implementation of ContextDataProvider.supplyStringMap which expects supplyContextData to return a mutable map.

But VertxContextDataProvider returns an immutable map and this is why sometimes (depending on log4j2 config) an UOE can be thrown.

In this commit, VertxContextDataProvider now returns a custom implementation of StringMap. This implementation is "frozen", which means log4j2 will not attempt to modify it.

Because log4j2 expects the StringMap entries to be processed in order, the ContextualDataImpl.getAll implementation has been modified to return a map sorted on its keys.
The map is not typed with SortedMap to reduce the number of methods to implement.